Add tests of status() function
This change was prompted by a PR from Filippo Biondi in which he added
tests of the status() function.  However, I prefer `Capture::Tiny` over
`Test::Trap`; also this change fixes a minor bug in the PR.  Beyond that
the test messages are a clearer description of what the tests are
attempting to do.